backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic==2.7.1
psycopg2-binary==2.9.9
databases==0.9.0
